MOB-113: extract Mob.Router and pin the hot-path property
MOB-112 gave Mob.Screen a third job. It was already the behaviour screens
implement and the macro generating their boilerplate; it became the process
owning navigation and every screen process too — nearly 1000 lines, with a
moduledoc that had to describe all three, which is part of why that moduledoc
has been wrong twice.

Mob.Router now holds navigation, the screen processes, and the :mob_screen
registered name. Mob.Screen keeps the behaviour and macro and delegates its
public API, so Mob.Screen.dispatch/3 and friends are unchanged. A move, not a
redesign: the process model landed in MOB-112 and is untouched.

The property MOB-113 actually exists to guarantee — the router must not be in
the per-message path — already held. MOB-111's listener delivers native events
straight to the owning screen's pid, so a tap goes native -> listener -> screen
-> sender with the router uninvolved.

What is new is that it is asserted rather than reasoned about.
router_hot_path_test.exs traces :receive on the router across a tap, a
value-carrying event, and a burst of fifty messages, and asserts the trace is
empty. Verified as a negative control: making the screen notify the router on
each message fails exactly those three tests.

That is worth a test rather than a comment. An earlier costing of this
architecture assumed a router in the loop and concluded per-screen processes
could not escape a hop per message; splitting the router from the sender is
what dissolved that, and a property that load-bearing should fail loudly when
someone breaks it.

Also: decode_file_result/3 stopped being a @doc false public function and
became private to Mob.Screen.Server, its only caller. And two guides were
making claims that were aspirational before MOB-112 and are now nearly true —
screen_lifecycle.md called each screen "a separate, supervised process", which
is half right: separate yes, supervised no, because the router restarts screens
itself since only it knows where a crashed one sat.

Rationale in decisions/2026-08-29-router-off-the-hot-path.md.

Tests: 5 new. Suite 1238 passed, format and credo --strict clean.
